format: 리뷰 페이지 내부 컴포넌트 설계 구조 개선
컴포넌트 외부에서 함수를 주입받아 사용해야 하던 드롭다운 메뉴의 동작을 페이지 단위에서 액션 함수를 지정할 수 있도록 컴포넌트 설계 수정
