+++ src/components/ReviewPage/ReviewContent/index.tsx
@@ -1,0 +1,18 @@
+import Tag from 'components/common/tag';
+import TAG from 'constants/tag';
+import * as S from './style';
+
+function ReviewContent({ content, tags }: { content: string; tags: (typeof TAG)[number]['id'][] }) {
+  return (
+    <>
+      <S.ReviewText>{content}</S.ReviewText>
+      <S.Tags>
+        {tags.map((tag) => (
+          <Tag id={tag} key={tag} />
+        ))}
+      </S.Tags>
+    </>
+  );
+}
+
+export default ReviewContent;
