+++ src/components/ReviewPage/ReviewContent/style.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const ReviewText = styled.p`
+  line-height: 1.8rem;
+  font-size: 1.2rem;
+  font-weight: 350;
+`;
+
+export const Tags = styled.div`
+  display: flex;
+  gap: 6px;
+  flex-wrap: wrap;
+  margin-top: 16px;
+`;
